+++ MicroBlink.framework/Headers/PPAusIDFrontRecognizerSettings.h
@@ -1,0 +1,63 @@
+//
+//  PPAusIDFrontRecognizerSettings.h
+//  BlinkIdFramework
+//
+//  Created by Dino on 30/06/16.
+//  Copyright © 2016 MicroBlink Ltd. All rights reserved.
+//
+
+#import "PPRecognizerSettings.h"
+
+/**
+ * Settings class for configuring Austrian Front ID Recognizer
+ *
+ * Austrian Front ID recognizer is used for scanning front side of Austrian ID.
+ */
+PP_CLASS_AVAILABLE_IOS(6.0) @interface PPAusIDFrontRecognizerSettings : PPRecognizerSettings
+
+/**
+ * Name of the image sent to didOutputMetadata method of scanDelegate object that contains face.
+ * This image will be sent to scan delegate during recognition process if displaying of face image
+ * is enabled via displayFaceImage property and receiving of dewarpedImage in MetadataSettings is enabled.
+ */
+@property (nonatomic, readonly) NSString* ID_FACE;
+
+/**
+ * Name of the image sent to didOutputMetadata method of scanDelegate object that contains signature.
+ * This image will be sent to scan delegate during recognition process if displaying of face image
+ * is enabled via displaySignatureImage property and receiving of dewarpedImage in MetadataSettings is enabled.
+ */
+@property (nonatomic, readonly) NSString* ID_SIGNATURE;
+
+/**
+ * Name of the image sent to didOutputMetadata method of scanDelegate object that contains full document.
+ * This image will be sent to scan delegate during recognition process if displaying of full document image
+ * is enabled via displayFullDocumentImage property and receiving of dewarpedImage in MetadataSettings is enabled.
+ */
+@property (nonatomic, readonly) NSString* FULL_DOCUMENT_IMAGE;
+
+/**
+ * Sets whether face photo from ID card should be sent to didOutputMetadata method of scanDelegate object.
+ * If you want to recieve this image, be sure to enable dewarpedImage in MetadataSettings.
+ *
+ * Default: NO
+ */
+@property (nonatomic) BOOL displayFacePhoto;
+
+/**
+ * Sets whether signature photo of ID card should be sent to didOutputMetadata method of scanDelegate object.
+ * If you want to recieve this image, be sure to enable dewarpedImage in MetadataSettings.
+ *
+ * Default: NO
+ */
+@property (nonatomic) BOOL displaySignaturePhoto;
+
+/**
+ * Sets whether full image of ID card should be sent to didOutputMetadata method of scanDelegate object.
+ * If you want to recieve this image, be sure to enable dewarpedImage in MetadataSettings.
+ *
+ * Default: NO
+ */
+@property (nonatomic) BOOL displayFullDocumentImage;
+
+@end