+++ MicroBlink.framework/Headers/PPBaseOverlayViewController.h
@@ -13,17 +13,17 @@
 NS_ASSUME_NONNULL_BEGIN
 
 /**
- * Common base class for PhotoPay default overlay view controllers
+ * Common base class for default overlay view controllers
  */
 PP_CLASS_AVAILABLE_IOS(6.0) @interface PPBaseOverlayViewController : PPOverlayViewController
 
 /**
- * Registers a subview for overlay events.
+ * Registers a subview for PPOverlaySubview protocol events.
  */
 - (void)registerOverlaySubview:(UIView<PPOverlaySubview>*)subview;
 
 /**
- * Unregisters a subview for overlay events.
+ * Unregisters a subview for PPOverlaySubview protocol events.
  */
 - (void)unregisterOverlaySubview:(UIView<PPOverlaySubview>*)subview;