+++ MicroBlink.framework/Headers/PPMultiDetectorSettings.h
@@ -9,10 +9,16 @@
 #import "PPDetectorSettings.h"
 
 /**
- * Settings for detector which combines multiple detectors inside
+ * Settings for detector which combines multiple detectors inside.
+ *
+ * As only one PPDetectorSettings can be used at a time, PPMultiDetectorSettings wraps multiple settings so they can be used simultaneously.
+ * In other words, it is necessary to use PPMultiDetectorSettings when more than 1 detectors are needed.
  */
 PP_CLASS_AVAILABLE_IOS(6.0) @interface PPMultiDetectorSettings : PPDetectorSettings
 
+/**
+ * Creates detector with array of concrete detectors.
+ */
 - (instancetype)initWithSettingsArray:(NSArray<__kindof PPDetectorSettings *> *)settingsArray;
 
 /**