+++ MicroBlink.framework/Headers/PPRegexOcrParserFactory.h
@@ -32,11 +32,11 @@ PP_CLASS_AVAILABLE_IOS(6.0) @interface PPRegexOcrParserFactory : PPOcrParserFact
 @property (nonatomic) NSString *regex;
 
 /**
- * Ocr Engine options used in Raw ocr parsing
+ * Sets desired options for a parser. Options enable various customizations of a parsers such as: allowed characters, character heigh, etc.
  *
  * Default: [[PPOcrEngineOptions alloc] init]
  */
-@property (nonatomic) PPOcrEngineOptions *options;
+-(void)setOptions:(PPOcrEngineOptions *) options;
 
 /**
  * Use algorithm for combining consecutive OCR results between video frames